done api vote start, api saved list

+++ Backend/src/main/java/hcmute/kltn/backend/entity/VoteStar.java
@@ -36,5 +36,5 @@ public class VoteStar {
     private User user;
 
     @Column(nullable = false)
-    private int star;
+    private float star;
 }
